+++ app/controllers/api/policies_controller.rb
@@ -57,25 +57,27 @@ module Api
       @json = JSON.parse(request.body.read)
       api_error(status: :bad_request) unless JSON::Validator.validate(schema, @json)
       return if check_existence
-      check_access
+      return if check_source
+      authorize_copy_move if copy_or_move_request?
     end
 
     def check_existence
       @resource = Resource.find_by(path: PathService.to_path(@json['path']))
       if @resource && copy_or_move_request?
-        resource_exist_api_error
+        api_error(status: :bad_request, errors: I18n.t('api.destination_resource_exists'))
         true
       else
         false
       end
     end
 
-    def resource_exist_api_error
-      api_error(status: :bad_request, errors: I18n.t('api.destination_resource_exists'))
-    end
-
-    def check_access
-      authorize_copy_move if copy_or_move_request?
+    def check_source
+      if copy_move_path && !Resource.find_by(path: PathService.to_path(copy_move_path))
+        api_error(status: :not_found, errors: I18n.t('api.source_policy_missing'))
+        true
+      else
+        false
+      end
     end
 
     def authorize_copy_move