+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ## [Unreleased]
 
 ### Added
+- Basic rack-attack configuration (@mkasztelnik)
 
 ### Changed
 - Update rubocop into 0.47.1, fix new discovered offenses (@mkasztelnik)